Changes are made on StudentUnitRecordManager.java
This class had issue on alot of code in one line , it have to break down to simplified version . This will be helpful for the other develop to understand and work. Some code contain more simplified approach. Also One comment done by the developer , is not clear. I would like to ask what did he ment by that.